time: turn Time.unix to i64, so it can represent times before 1970-01-01, fix time operators, add more tests (#11050)

+++ vlib/rand/rand.v
@@ -275,7 +275,7 @@ const (
 // users or business transactions.
 // (https://news.ycombinator.com/item?id=14526173)
 pub fn ulid() string {
-	return ulid_at_millisecond(time.utc().unix_time_milli())
+	return ulid_at_millisecond(u64(time.utc().unix_time_milli()))
 }
 
 // ulid_at_millisecond does the same as `ulid` but takes a custom Unix millisecond timestamp via `unix_time_milli`.
